+++ packages/plugin-gallery/web/src/components/layout-controls-section.js
@@ -15,6 +15,11 @@ import {
 import ImageRatioSelector from './gallery-controls/image-ratio-selector';
 import ThumbnailPlacementSelector from './gallery-controls/thumbnail-placement-selector';
 
+const scrollDirectionOptions = {
+  horizontal: { oneRow: true, showArrows: true, isVertical: false },
+  vertical: { oneRow: false, showArrows: false },
+};
+
 class Separator extends Component {
   static propTypes = {
     type: PropTypes.string.isRequired,
@@ -136,7 +141,7 @@ class LayoutControlsSection extends Component {
       component: ScrollDirection,
       props: {
         onChange: value => {
-          return this.applyGallerySetting({ oneRow: value === 'horizontal', isVertical: false });
+          return this.applyGallerySetting(scrollDirectionOptions[value]);
         },
         value: this.getValueFromComponentStyles('oneRow') ? 'horizontal' : 'vertical',
         t,